+++ src/main/java/no/ntnu/okse/protocol/amqp/AMQPServer.java
@@ -24,6 +24,9 @@
 
 package no.ntnu.okse.protocol.amqp;
 
+import no.ntnu.okse.core.messaging.MessageService;
+import no.ntnu.okse.core.topic.Topic;
+import no.ntnu.okse.core.topic.TopicService;
 import org.apache.log4j.Logger;
 import org.apache.qpid.proton.amqp.messaging.Accepted;
 import org.apache.qpid.proton.engine.BaseHandler;
@@ -32,6 +35,8 @@ import org.apache.qpid.proton.engine.Event;
 import org.apache.qpid.proton.engine.Link;
 import org.apache.qpid.proton.engine.Receiver;
 import org.apache.qpid.proton.engine.Sender;
+import org.apache.qpid.proton.message.Message;
+import org.apache.qpid.proton.messenger.impl.Address;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
@@ -153,4 +158,45 @@ public class AMQPServer extends BaseHandler {
 //        }
 //    }
 
+    @Override
+    public void onDelivery(Event event) {
+        log.debug("I got a delivery");
+        Delivery dlv = event.getDelivery();
+        Link link = dlv.getLink();
+        if (link instanceof Sender) {
+            dlv.settle();
+        } else {
+            Receiver rcv = (Receiver) link;
+            if (!dlv.isPartial()) {
+                byte[] bytes = new byte[dlv.pending()];
+                rcv.recv(bytes, 0, bytes.length);
+                Message msg = Message.Factory.create();
+                msg.decode(bytes, 0, bytes.length);
+                Address address = new Address(msg.getAddress());
+
+                TopicService.getInstance().addTopic(address.getName());
+
+                Topic t = TopicService.getInstance().getTopic(address.getName());
+
+                no.ntnu.okse.core.messaging.Message message =
+                        new no.ntnu.okse.core.messaging.Message(msg.getBody().toString(), t, null);
+                message.setOriginProtocol(protocolServerType);
+                
+                MessageService.getInstance().distributeMessage(message);
+                totalMessages++;
+
+                System.out.println(address.getName());
+                System.out.println(msg.getBody().toString());
+                //String address = router.getAddress(rcv);
+//                MessageBytes messageBytes = new MessageBytes(bytes);
+//                messages.put(address, messageBytes);
+//                dlv.disposition(Accepted.getInstance());
+//                dlv.settle();
+//                if (!quiet) {
+//                    log.debug(String.format("Got message(%s): %s", address, messageBytes));
+//                }
+//                send(address);
+            }
+        }
+    }
 }
